refactor: rename runcat-statusline.py to statusline.py
