+++ .config/claude/settings.json
@@ -103,7 +103,7 @@
   },
   "statusLine": {
     "type": "command",
-    "command": "if [ -x \"$HOME/.claude/runcat-statusline.py\" ]; then \"$HOME/.claude/runcat-statusline.py\"; else ccstatusline; fi"
+    "command": "\"$HOME/.claude/statusline.py\""
   },
   "enabledPlugins": {
     "rust-analyzer-lsp@claude-plugins-official": true,
