+++ conftest.py
@@ -41,6 +41,6 @@ _load_module(
 _gh_ops.append_review = sys.modules["gh_ops.append_review"]
 
 _load_module(
-    "config.claude.runcat_statusline",
-    _PROJECT_ROOT / ".config" / "claude" / "runcat-statusline.py",
+    "config.claude.statusline",
+    _PROJECT_ROOT / ".config" / "claude" / "statusline.py",
 )
